add wedgeAbsFraction... viewer to the right
The original page plotted only the data in
`wedgeFractionObstructedList`. The data in my
`wedgeAbsFractionObstructionList` does not correlate in an obvious way,
to me, so I added a second viewer to plot that data as well.

The addition is quick and dirty, just to be able to see them side by
side.